pyproject.toml
[tool.poetry]
name = "security-rag"
version = "0.1.0"
description = "Security RAG for LLM basic course (ITMO)"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "1.26.4"
pandas = "^2.2.3"
matplotlib = "^3.9.3"
langchain = "^0.3.10"
fastapi = "^0.115.6"
python-dotenv = "^1.0.1"
langchain-huggingface = "^0.1.2"
chromadb = "^0.5.23"
langchain-mistralai = "^0.2.3"
langchain-community = "^0.3.10"
langchain-chroma = "^0.1.4"
datasets = "^3.2.0"
langsmith = "^0.2.3"
slowapi = "^0.1.9"
torch = "2.5.1"
langchain-ollama = "^0.2.2"
aiogram = "^3.16.0"
langfuse = "^2.57.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
